This commit changes RicherString to take advantage of the implicit classes and value classes added in Scala 2.10.  RicherString is now an implicit value class.  These are not allowed by the compiler to be top-level classes, so it was moved into the org.scalatra.util package object. Users of RicherString no longer need to explicitly import it, instead just import the package object.  The advantage to this is that the compiler can eliminate the wrapper class and basically do a method call on a static object which is faster than instantiating the wrapper and calling a method on it.
