+++ atmosphere/src/main/scala/org/scalatra/atmosphere/AtmosphereSupport.scala
@@ -19,7 +19,7 @@ import org.jboss.servlet.http.{ HttpEvent, HttpEventServlet }
 import org.json4s._
 import org.scalatra.json.JsonSupport
 import org.scalatra.servlet.ScalatraAsyncSupport
-import org.scalatra.util.RicherString._
+import org.scalatra.util._
 
 import scala.collection.JavaConverters._
 import scala.util.control.Exception.allCatch
