+++ commands/src/main/scala/org/scalatra/commands/field.scala
@@ -2,7 +2,7 @@ package org.scalatra
 package commands
 
 import mojolly.inflector.InflectorImports._
-import org.scalatra.util.RicherString._
+import org.scalatra.util._
 import org.scalatra.util.conversion._
 import org.scalatra.validation._
