+++ slf4j/src/main/scala/org/scalatra/slf4j/LevelColorizer.scala
@@ -4,7 +4,7 @@ package slf4j
 import ch.qos.logback.classic.Level
 import ch.qos.logback.classic.pattern.ClassicConverter
 import ch.qos.logback.classic.spi.ILoggingEvent
-import org.scalatra.util.RicherString._
+import org.scalatra.util._
 
 object LevelColorizer {
   private val EndColor = "\u001b[m"
